Move Config class to root
