+++ src/Core/src/Internal/Constructor.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Spiral\Core\Internal;
 
+use Spiral\Core\Config;
+
 /**
  * @internal
  */
